server/pyproject.toml
[project]
name = "multivox"
version = "0.1.0"
description = "Generative dialogue system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.8",
    "genanki>=0.13.1",
    "google-cloud-texttospeech>=2.25.0",
    "google-genai>=1.0.0",
    "ipykernel>=6.29.5",
    "litellm>=1.60.7",
    "mypy>=1.15.0",
    "pandas>=2.2.3",
    "playwright>=1.50.0",
    "pydantic-settings>=2.7.1",
    "pysrt>=1.1.2",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "pytest-repeat>=0.9.3",
    "pytest-xdist>=3.6.1",
    "silero-vad>=5.1.2",
    "soundfile>=0.13.1",
    "typer>=0.15.1",
    "uvicorn>=0.34.0",
    "websockets>=14.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["multivox"]

[dependency-groups]
dev = [
    "hatch>=1.14.0",
]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
addopts = "-v --asyncio-mode=auto --log-cli-level=INFO --log-cli-format='%(asctime)s %(levelname)s %(message)s'"
